chore(daemon): distribute tool impls
