+++ crates/daemon/src/hook/skill/mod.rs
@@ -13,3 +13,4 @@ pub use {
 mod handler;
 pub mod loader;
 pub mod registry;
+pub(crate) mod tool;
